Module amazon_sp_aplus_content::apis::aplus_content_api

Enums

CreateContentDocumentError

struct for typed errors of method create_content_document

GetContentDocumentError

struct for typed errors of method get_content_document

ListContentDocumentAsinRelationsError

struct for typed errors of method list_content_document_asin_relations

PostContentDocumentApprovalSubmissionError

struct for typed errors of method post_content_document_approval_submission

PostContentDocumentAsinRelationsError

struct for typed errors of method post_content_document_asin_relations

PostContentDocumentSuspendSubmissionError

struct for typed errors of method post_content_document_suspend_submission

SearchContentDocumentsError

struct for typed errors of method search_content_documents

SearchContentPublishRecordsError

struct for typed errors of method search_content_publish_records

UpdateContentDocumentError

struct for typed errors of method update_content_document

ValidateContentDocumentAsinRelationsError

struct for typed errors of method validate_content_document_asin_relations

Functions

create_content_document

Creates a new A+ Content document. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_content_document

Returns an A+ Content document, if available. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

list_content_document_asin_relations

Returns a list of ASINs related to the specified A+ Content document, if available. If you do not include the asinSet parameter, the operation returns all ASINs related to the content document. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

post_content_document_approval_submission

Submits an A+ Content document for review, approval, and publishing. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

post_content_document_asin_relations

Replaces all ASINs related to the specified A+ Content document, if available. This may add or remove ASINs, depending on the current set of related ASINs. Removing an ASIN has the side effect of suspending the content document from that ASIN. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

post_content_document_suspend_submission

Submits a request to suspend visible A+ Content. This neither deletes the content document nor the ASIN relations. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

search_content_documents

Returns a list of all A+ Content documents assigned to a selling partner. This operation returns only the metadata of the A+ Content documents. Call the getContentDocument operation to get the actual contents of the A+ Content documents. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

search_content_publish_records

Searches for A+ Content publishing records, if available. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

update_content_document

Updates an existing A+ Content document. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

validate_content_document_asin_relations

Checks if the A+ Content document is valid for use on a set of ASINs. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.