Function amazon_sp_aplus_content::apis::aplus_content_api::update_content_document

pub async fn update_content_document(
    configuration: &Configuration,
    content_reference_key: &str,
    marketplace_id: &str,
    post_content_document_request: PostContentDocumentRequest
) -> Result<PostContentDocumentResponse, Error<UpdateContentDocumentError>>

Updates an existing A+ Content document. Usage Plans: | Plan type | Rate (requests per second) | Burst | | –– | –– | –– | |Default| 10 | 10 | |Selling partner specific| Variable | Variable | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation. Rate limits for some selling partners will vary from the default rate and burst shown in the table above. For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.